Documentation:
- GXMat4 page has been updated
- GXVec3 page has been updated
- VVL compilation page has been updated
- shader compilation page has been updated
- BitField page has been added
- Lua API page has been refactored and extended
- build requirements have been updated

Project:
- DirectX Shader Compiler has been updated to 1.7.2308.10184, ea4ff2972b32fe005e91ddb336b2471d6ab3bdd6
- SPIR-V shader blobs have been recompiled (no changes)
- FreeType has been updated to 2.13.2, 8c5ec8dd63ffd3f76ba0531dda15d2993ab0f20b
- VVL has been updated to 0117636eacfae8ae978b34618dd9f37614efd973

3rd-party:
- float32_t and int32_t as internal types of Lua
- FreeType headers have been updated
- FreeType binary has been updated
- VVL binary has been updated

android-vulkan:
- likely|unlikely attributes
- code style

pbr:
- likely|unlikely attributes
- bit field entity
- Scene handlers and float instead of double for delta time
- BitField as collision group type for Lua API
- Lua stack safety check for GXMat3, GXQuat, GXVec3 and GXVec4

+++ app/src/main/cpp/include/pbr/scene.hpp
@@ -35,7 +35,7 @@ class Scene final
         RenderSession*                                  _renderSession = nullptr;
         UILayerList                                     _uiLayerList {};
 
-        lua_Number                                      _aspectRatio = 1.0;
+        lua_Number                                      _aspectRatio = 1.0F;
         lua_Integer                                     _width = -1;
         lua_Integer                                     _height = -1;
 
@@ -69,7 +69,7 @@ class Scene final
         [[nodiscard]] GXMat4 const &GetActiveCameraProjectionMatrix () const noexcept;
         [[nodiscard]] android_vulkan::Physics &GetPhysics () noexcept;
 
-        [[nodiscard]] bool OnAnimationUpdated ( double deltaTime ) noexcept;
+        [[nodiscard]] bool OnAnimationUpdated ( float deltaTime ) noexcept;
 
         [[nodiscard]] bool OnInitDevice ( android_vulkan::Renderer &renderer,
             RenderSession &renderSession,
@@ -84,10 +84,10 @@ class Scene final
         void OnPause () noexcept;
         void OnResume () noexcept;
 
-        [[nodiscard]] bool OnPrePhysics ( double deltaTime ) noexcept;
-        [[nodiscard]] bool OnPostPhysics ( double deltaTime ) noexcept;
-        [[nodiscard]] bool OnResolutionChanged ( VkExtent2D const &resolution, double aspectRatio ) noexcept;
-        [[nodiscard]] bool OnUpdate ( double deltaTime ) noexcept;
+        [[nodiscard]] bool OnPrePhysics ( float deltaTime ) noexcept;
+        [[nodiscard]] bool OnPostPhysics ( float deltaTime ) noexcept;
+        [[nodiscard]] bool OnResolutionChanged ( VkExtent2D const &resolution, float aspectRatio ) noexcept;
+        [[nodiscard]] bool OnUpdate ( float deltaTime ) noexcept;
 
         [[nodiscard]] bool LoadScene ( android_vulkan::Renderer &renderer,
             char const* scene,
@@ -97,7 +97,7 @@ class Scene final
         void RemoveActor ( Actor const &actor ) noexcept;
         void Submit ( android_vulkan::Renderer &renderer ) noexcept;
 
-        static void OnUpdateAnimations ( double deltaTime, size_t commandBufferIndex ) noexcept;
+        static void OnUpdateAnimations ( float deltaTime, size_t commandBufferIndex ) noexcept;
 
     private:
         void AppendActor ( ActorRef &actor ) noexcept;
